// src/models/User.ts
export interface User {
    nombre: string;
    apellidos: string;
    email: string;
    edad: number;
    direccion: string;
    nameTutor: string;
    ageTutor: number;
    poblacion: string;
    provincia: string;
    codPostal: number;
    telefonoFijo: number;
    movil: number;
    fechaNacimiento: string;
    nif: string;
    image?: File
}